Problem: Sometimes we need to hold for post-mortems a resource automatically reserved by a job (#142)
Solution: Add a button to "steal" an already reserved resource.

Note that this solution is relatively naive, doing unreserve+reserve
(so there is technically a time-gap that some other queued consumer
might grab this resource; hopefully it is disregardably short).

Also the build-vs-user matching might better use regexes ('#[0-9]+$')
but I did not quickly pick how.

If needed, this can surely be improved. For now consider an MVP,
usable but rough-edged ;)

Signed-off-by: Jim Klimov <EvgenyKlimov@eaton.com>